backend/requirements.txt
fastapi
uvicorn
requests
pandas
numpy
scikit-learn
gdown
python-dotenv